Refine RamFs by extracting immutable metadata from the lock

+++ kernel/aster-nix/src/fs/ramfs/mod.rs
@@ -8,5 +8,5 @@ mod fs;
 
 const RAMFS_MAGIC: u64 = 0x0102_1994;
 const BLOCK_SIZE: usize = 4096;
-const ROOT_INO: usize = 1;
+const ROOT_INO: u64 = 1;
 const NAME_MAX: usize = 255;
